- Se cambiaron los tipo de datos de algunas entidades tipo int a (double)
- Se arreglaron los ComboBox de EntradasLibros y SalidasLibros

- Se pusieron en comentario las ForeignKey de UsuarioId, porque no se estaban utilizando

- Se agregó una nueva Migracion

- Cambios menores en general

+++ Entidades/Estudiantes.cs
@@ -16,13 +16,13 @@ namespace ProyectoFinal_PrestamosLibros.Entidades
         //[ForeignKey("UsuarioId")]
         //public Usuarios Usuario { get; set; } = new Usuarios();
         //————————————————————————————————————————————————————————————————————————————
-        public int Matricula { get; set; }
+        public double Matricula { get; set; }
         public String Nombres { get; set; }
         public String Apellidos { get; set; }
         public DateTime FechaNacimiento { get; set; } = DateTime.Now;
-        public int Cedula { get; set; }
+        public double Cedula { get; set; }
         public String Genero { get; set; }
-        public int Telefono { get; set; }
+        public double Telefono { get; set; }
         public String Direccion { get; set; }
         public String Correo { get; set; }
     }
